feat/fix: add user service; fix user model
