+++ src/user/user.controller.ts
@@ -3,7 +3,6 @@ import {
   Res,
   Param,
   Body,
-  HttpStatus,
   Get,
   Post,
   Put,
@@ -11,57 +10,26 @@ import {
 } from '@nestjs/common';
 import { Response } from 'express';
 
-import db from '../db/db';
 import { ICreateUserDto, IUpdatePasswordDto } from 'src/types/types';
+import { UserService } from './user.service';
 
 @Controller('user')
 export class UserController {
+  constructor(private readonly userService: UserService) {}
+
   @Get()
   getAll() {
-    return db.users.findAll();
+    return this.userService.getAll();
   }
 
   @Get(':id')
   getById(@Param('id') id: string, @Res() response: Response) {
-    const getUserByIdResult = db.users.findById(id);
-    switch (getUserByIdResult) {
-      case 'invalid uuid':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'invalid uuid' });
-        break;
-      case "entity doesn't exist":
-        response
-          .status(HttpStatus.NOT_FOUND)
-          .send({ message: 'the user with the specified ID was not found' });
-        break;
-      default:
-        response.status(HttpStatus.OK).send(getUserByIdResult);
-    }
+    return this.userService.getById(id, response);
   }
 
   @Post()
   createUser(@Body() createUserDto: ICreateUserDto, @Res() response: Response) {
-    const userCreationResult = db.users.create(createUserDto);
-    switch (userCreationResult) {
-      case 'insufficient data for creation':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'insufficient data to create user' });
-        break;
-      case 'login already in use':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'specified login name is already in use' });
-        break;
-      case 'invalid data':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'invalid data received' });
-        break;
-      default:
-        response.status(HttpStatus.CREATED).send(userCreationResult);
-    }
+    return this.userService.createUser(createUserDto, response);
   }
 
   @Put(':id')
@@ -70,49 +38,11 @@ export class UserController {
     @Body() updatePasswordDto: IUpdatePasswordDto,
     @Res() response: Response,
   ) {
-    const userUpdatingResult = db.users.updatePassword(id, updatePasswordDto);
-    switch (userUpdatingResult) {
-      case 'invalid uuid':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'invalid uuid' });
-        break;
-      case "passwords don't match":
-        response
-          .status(HttpStatus.FORBIDDEN)
-          .send({ message: 'the specified old password is wrong' });
-        break;
-      case "user doesn't exist":
-        response
-          .status(HttpStatus.NOT_FOUND)
-          .send({ message: 'the user with the specified ID was not found' });
-        break;
-      case 'invalid data':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'invalid data received' });
-        break;
-      default:
-        response.status(HttpStatus.OK).send(userUpdatingResult);
-    }
+    return this.userService.updateUser(id, updatePasswordDto, response);
   }
 
   @Delete(':id')
   deleteUser(@Param('id') id: string, @Res() response: Response) {
-    const userDeletionResult = db.users.delete(id);
-    switch (userDeletionResult) {
-      case 'invalid uuid':
-        response
-          .status(HttpStatus.BAD_REQUEST)
-          .send({ message: 'invalid uuid' });
-        break;
-      case "entity doesn't exist":
-        response
-          .status(HttpStatus.NOT_FOUND)
-          .send({ message: 'the user with the specified ID was not found' });
-        break;
-      case 'success':
-        response.status(HttpStatus.NO_CONTENT).send();
-    }
+    return this.userService.deleteUser(id, response);
   }
 }
